requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
# 2.0.36+ ломается на Python 3.14 (typing.Union / PEP 649) — нужен свежий 2.0.x
sqlalchemy==2.0.48
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0
